task 3.1.1

+++ src/main/java/com/example/springbootdemo/service/UserService.java
@@ -5,12 +5,12 @@ import com.example.springbootdemo.model.User;
 import java.util.List;
 
 public interface UserService {
-    public User findBuId(Long id);
+    public User findBuId(int id);
 
     public List<User> findAll();
 
     public User saveUser(User user);
 
-    public void deleteById(Long id);
+    public void deleteById(int id);
 
 }
